Do not delete TicketLayout backgrounds long as they're referenced somewhere

+++ src/pretix/plugins/badges/views.py
@@ -227,7 +227,7 @@ class LayoutEditorView(BaseEditorView):
         return self.layout.background.url if self.layout.background else self.get_default_background()
 
     def save_background(self, f: CachedFile):
-        if self.layout.background:
+        if self.layout.background and BadgeLayout.objects.filter(background=self.layout.background).count() == 1:
             self.layout.background.delete()
         self.layout.background.save('background.pdf', f.file)
